Trials mismatch
I think there is a problem between the N of Clean trials and the N of trials with actual markers. It is either a problem that some trials do not have markers, or have other markers? Anyway, even for IN-1107 that is completely fine, we have 55 = FastERP_info.N_trials.FaceUp.Nclean, but only 33 = length(EEGdata_Faces_Obj.trialinfo)

For now I have changed length(EEGdata_Faces_Obj.trialinfo) > 10, line 61 of thee Individual_GrandAverages.m, and it runs, but we should understand what is going on.