Running prep-data & clean-data on the analysis sample
The scripts run, but there are unidentified bugs that stop the script. This is probably due to problems with the data (see below) but the scripts should be improved so that the problem is logged, rather than the battery to stop. The problems for this to happen are: tNev_EEG < 10, unknown, and missing/mismatching info file (no error/warning, but n of clean trials entered as = 999). Current strategy is to exclude these participants without debugging.